+++ topobank/users/tests/test_prevent_url_guessing.py
@@ -1,8 +1,9 @@
-
 from django.urls import reverse
 from django.contrib.auth.models import Permission
 
-def test_prevent_username_guessing(client, django_user_model):
+from .test_utils import are_collaborating, SurfaceFactory
+
+def test_sharing_profile(client, django_user_model):
 
     user1 = django_user_model.objects.create_user(username='testuser1', password="abcd$1234")
     user2 = django_user_model.objects.create_user(username='testuser2', password="abcd$5678")
@@ -17,8 +18,21 @@ def test_prevent_username_guessing(client, django_user_model):
     response = client.get(reverse('users:detail', kwargs={'username': 'testuser1'}))  # same user
     assert response.status_code == 200
 
+    #
+    # both users don't share anything, so they can't see each others profiles
+    #
+    assert not are_collaborating(user1, user2)
+
     response = client.get(reverse('users:detail', kwargs={'username': 'testuser2'})) # other user!!
     assert response.status_code == 403 # Forbidden
+
+    # share sth. and the view of profile is possible
+    surface = SurfaceFactory(user=user1)
+    surface.share(user2)
+
+    assert are_collaborating(user1, user2)
+    response = client.get(reverse('users:detail', kwargs={'username': 'testuser2'}))
+    assert response.status_code == 200  # Allowed